pyproject.toml
[tool.poetry]
name = "bica-stones"
version = "0.1.0"
description = "dependencies for telegram bot"
authors = ["Artem Kostyaev <a.s.kostyaev@tinkoff.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.13.1"
requests = "2.32.3"
fastapi = "0.115.0"
python-dotenv = "1.0.1"
aiofiles = "24.1.0"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.6.0"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
click = "8.1.7"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi-cli = "0.0.5"
frozenlist = "1.4.1"
h11 = "0.14.0"
httpcore = "1.0.6"
httptools = "0.6.1"
httpx = "0.27.2"
idna = "3.10"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
magic-filter = "1.0.12"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
multidict = "6.1.0"
orjson = "3.10.7"
propcache = "0.2.0"
psycopg = "3.2.3"
psycopg-binary = "3.2.3"
psycopg-c = "3.2.3"
psycopg-pool = "3.2.3"
pydantic = "2.9.2"
pydantic-extra-types = "2.9.0"
pydantic-settings = "2.5.2"
pydantic-core = "2.23.4"
pygments = "2.18.0"
python-multipart = "0.0.12"
pyyaml = "6.0.2"
rich = "13.9.2"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.38.6"
typer = "0.12.5"
typing-extensions = "4.12.2"
ujson = "5.10.0"
urllib3 = "2.2.3"
uvicorn = "0.31.0"
uvloop = "0.20.0"
watchfiles = "0.24.0"
websockets = "13.1"
yarl = "1.15.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
